Add hibernate repo classes
