+++ src/main/java/ru/maynim/astonmvc/controller/FileController.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import ru.maynim.astonmvc.model.File;
+import ru.maynim.astonmvc.entity.File;
 import ru.maynim.astonmvc.repository.FileRepository;
 import ru.maynim.astonmvc.repository.NoteRepository;
