+++ src/main/java/ru/maynim/astonmvc/entity/Note.java
@@ -1,4 +1,4 @@
-package ru.maynim.astonmvc.model;
+package ru.maynim.astonmvc.entity;
 
 import lombok.*;
 
@@ -9,9 +9,9 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class File {
+public class Note {
     private long id;
     private String name;
-    private String url;
-    private Note note;
+    private String content;
+    private User user;
 }
